bacula/src/plugins/fd/kubernetes-backend/requirements.txt
# -*- coding: UTF-8 -*-
#
# Copyright (C) 2000-2020 Kern Sibbald
# License: BSD 2-Clause; see file LICENSE-FOSS
#
# Author: Radoslaw Korzeniewski
#

pyyaml == 5.3.1
kubernetes <= 11
urllib3 == 1.26
requests == 2.25
